HSLdevcom/bultti#204 Refactor table component into multiple modules.
