+++ src/common/table/FilteredResponseTable.tsx
@@ -4,9 +4,9 @@ import { observer } from 'mobx-react-lite'
 import { LoadingDisplay } from '../components/Loading'
 import TableFiltersControl from './TableFiltersControl'
 import { TableStateType } from './useTableState'
-import { IFilteredSortedResponse, useRenderCellValue } from './tableUtils'
+import { CellValType, IFilteredSortedResponse, useRenderCellValue } from './tableUtils'
 import { pick } from 'lodash'
-import { CellValType, TablePropTypes } from './Table'
+import { TablePropTypes } from './Table'
 import { EmptyView } from '../components/Messages'
 import { Text } from '../../util/translate'
 import PagedTable from './PagedTable'
