+++ src/contract/ContractEditor.tsx
@@ -29,7 +29,6 @@ import { FlexRow } from '../common/components/common'
 import FileUploadInput from '../common/input/FileUploadInput'
 import { useUploader } from '../util/useUploader'
 import { SubHeading } from '../common/components/Typography'
-import Table from '../common/table/Table'
 import { text, Text } from '../util/translate'
 import { navigateWithQueryString } from '../util/urlValue'
 import PagedTable from '../common/table/PagedTable'
