+++ src/executionRequirement/ObservedRequirementsTable.tsx
@@ -1,12 +1,13 @@
 import React, { useCallback, useMemo } from 'react'
 import styled from 'styled-components/macro'
 import { observer } from 'mobx-react-lite'
-import Table, { EditValue, TableEditProps } from '../common/table/Table'
+import Table from '../common/table/Table'
 import { isNumeric } from '../util/isNumeric'
 import { ObservedExecutionValue } from '../schema-types'
 import { orderBy } from 'lodash'
 import { getTotal } from '../util/getTotal'
 import { round } from '../util/round'
+import { EditValue, TableEditProps } from '../common/table/tableUtils'
 
 const ExecutionRequirementsAreaContainer = styled.div`
   margin-top: 1.5rem;
